Add server side auth
Auth on the server makes more sense and is closer to what one would expect from
a single page app. Unfortunately, it adds the complexity of running a server
and client app with `npm run server` in one tab and `npm run start:dev` in
another. Added tests for Logout.js